Migrate ResourceIcon to the SDK
