+++ frontend/packages/console-dynamic-plugin-sdk/src/app/components/index.ts
@@ -5,6 +5,7 @@ export { default as PopoverStatus } from './status/PopoverStatus';
 export { default as StatusComponent } from './status/Status';
 export { default as StatusIconAndText } from './status/StatusIconAndText';
 export { default as ResourceStatus } from './utils/resource-status';
+export { ResourceIcon, ResourceIconProps } from './utils/resource-icon';
 
 export * from './status/icons';
 export * from './status/statuses';
