+++ frontend/packages/console-dynamic-plugin-sdk/src/utils/k8s/index.ts
@@ -1,3 +1,4 @@
+export * from './k8s-get-resource';
 export * from './k8s-resource';
 export * from './k8s-utils';
 export * from './k8s-ref';
